requirement.txt
pandas==1.5.3
Pillow==9.4.0
torch==2.1.0
torchvision==0.15.0
torchtext==0.15.0
spacy==3.6.0
scikit-learn==1.2.0
tqdm==4.65.0
